Struct libxslt::bindings::_xmlParserCtxt

#[repr(C)]

pub struct _xmlParserCtxt {
    pub sax: *mut _xmlSAXHandler,
    pub userData: *mut c_void,
    pub myDoc: *mut _xmlDoc,
    pub wellFormed: i32,
    pub replaceEntities: i32,
    pub version: *const u8,
    pub encoding: *const u8,
    pub standalone: i32,
    pub html: i32,
    pub input: *mut _xmlParserInput,
    pub inputNr: i32,
    pub inputMax: i32,
    pub inputTab: *mut *mut _xmlParserInput,
    pub node: *mut _xmlNode,
    pub nodeNr: i32,
    pub nodeMax: i32,
    pub nodeTab: *mut *mut _xmlNode,
    pub record_info: i32,
    pub node_seq: _xmlParserNodeInfoSeq,
    pub errNo: i32,
    pub hasExternalSubset: i32,
    pub hasPErefs: i32,
    pub external: i32,
    pub valid: i32,
    pub validate: i32,
    pub vctxt: _xmlValidCtxt,
    pub instate: i32,
    pub token: i32,
    pub directory: *mut i8,
    pub name: *const u8,
    pub nameNr: i32,
    pub nameMax: i32,
    pub nameTab: *mut *const u8,
    pub nbChars: i64,
    pub checkIndex: i64,
    pub keepBlanks: i32,
    pub disableSAX: i32,
    pub inSubset: i32,
    pub intSubName: *const u8,
    pub extSubURI: *mut u8,
    pub extSubSystem: *mut u8,
    pub space: *mut i32,
    pub spaceNr: i32,
    pub spaceMax: i32,
    pub spaceTab: *mut i32,
    pub depth: i32,
    pub entity: *mut _xmlParserInput,
    pub charset: i32,
    pub nodelen: i32,
    pub nodemem: i32,
    pub pedantic: i32,
    pub _private: *mut c_void,
    pub loadsubset: i32,
    pub linenumbers: i32,
    pub catalogs: *mut c_void,
    pub recovery: i32,
    pub progressive: i32,
    pub dict: *mut _xmlDict,
    pub atts: *mut *const u8,
    pub maxatts: i32,
    pub docdict: i32,
    pub str_xml: *const u8,
    pub str_xmlns: *const u8,
    pub str_xml_ns: *const u8,
    pub sax2: i32,
    pub nsNr: i32,
    pub nsMax: i32,
    pub nsTab: *mut *const u8,
    pub attallocs: *mut i32,
    pub pushTab: *mut *mut c_void,
    pub attsDefault: *mut _xmlHashTable,
    pub attsSpecial: *mut _xmlHashTable,
    pub nsWellFormed: i32,
    pub options: i32,
    pub dictNames: i32,
    pub freeElemsNr: i32,
    pub freeElems: *mut _xmlNode,
    pub freeAttrsNr: i32,
    pub freeAttrs: *mut _xmlAttr,
    pub lastError: _xmlError,
    pub parseMode: u32,
    pub nbentities: u64,
    pub sizeentities: u64,
    pub nodeInfo: *mut _xmlParserNodeInfo,
    pub nodeInfoNr: i32,
    pub nodeInfoMax: i32,
    pub nodeInfoTab: *mut _xmlParserNodeInfo,
    pub input_id: i32,
    pub sizeentcopy: u64,
}

xmlParserCtxt:

The parser context. NOTE This doesn't completely define the parser state, the (current ?) design of the parser uses recursive function calls since this allow and easy mapping from the production rules of the specification to the actual code. The drawback is that the actual function call also reflect the parser state. However most of the parsing routines takes as the only argument the parser context pointer, so migrating to a state based parser for progressive parsing shouldn't be too hard.

Fields
sax: *mut _xmlSAXHandler userData: *mut c_void myDoc: *mut _xmlDoc wellFormed: i32 replaceEntities: i32 version: *const u8 encoding: *const u8 standalone: i32 html: i32 input: *mut _xmlParserInput inputNr: i32 inputMax: i32 inputTab: *mut *mut _xmlParserInput node: *mut _xmlNode nodeNr: i32 nodeMax: i32 nodeTab: *mut *mut _xmlNode record_info: i32 node_seq: _xmlParserNodeInfoSeq errNo: i32 hasExternalSubset: i32 hasPErefs: i32 external: i32 valid: i32 validate: i32 vctxt: _xmlValidCtxt instate: i32 token: i32 directory: *mut i8 name: *const u8 nameNr: i32 nameMax: i32 nameTab: *mut *const u8 nbChars: i64 checkIndex: i64 keepBlanks: i32 disableSAX: i32 inSubset: i32 intSubName: *const u8 extSubURI: *mut u8 extSubSystem: *mut u8 space: *mut i32 spaceNr: i32 spaceMax: i32 spaceTab: *mut i32 depth: i32 entity: *mut _xmlParserInput charset: i32 nodelen: i32 nodemem: i32 pedantic: i32 _private: *mut c_void loadsubset: i32 linenumbers: i32 catalogs: *mut c_void recovery: i32 progressive: i32 dict: *mut _xmlDict atts: *mut *const u8 maxatts: i32 docdict: i32 str_xml: *const u8 str_xmlns: *const u8 str_xml_ns: *const u8 sax2: i32 nsNr: i32 nsMax: i32 nsTab: *mut *const u8 attallocs: *mut i32 pushTab: *mut *mut c_void attsDefault: *mut _xmlHashTable attsSpecial: *mut _xmlHashTable nsWellFormed: i32 options: i32 dictNames: i32 freeElemsNr: i32 freeElems: *mut _xmlNode freeAttrsNr: i32 freeAttrs: *mut _xmlAttr lastError: _xmlError parseMode: u32 nbentities: u64 sizeentities: u64 nodeInfo: *mut _xmlParserNodeInfo nodeInfoNr: i32 nodeInfoMax: i32 nodeInfoTab: *mut _xmlParserNodeInfo input_id: i32 sizeentcopy: u64